+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Refactoring
 - Fix copy from list
 - Fix editor in Safari
+- Working in ILIAS tests
 
 ## [1.0.0]
 - First version